[YYB][RTOS][BOOTLOADER] Boot app after bootloader.
Symptom:
Root Cause: N/A
Solution: N/A
Project: YYB.
Note: On startup the bootloader checks the bootloader settings page if
the valid application flag is set or not. If it is set, then the
bootloader will branch to the the application. If it is not set, then
the device will stay in bootloader mode and wait for an application DFU.
Test done by RD
   -Build pass: Y
   -Feature verification: Y